refactor: move tests utilities to a folder called __helpers__
