+++ src/__tests__/__helpers__/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './tests'
-export * from './mocks'
 export * from './headless-ui'
 export * from './base64-to-blob'
