+++ src/__tests__/app/admin/components/add-link-form/index.test.tsx
@@ -1,4 +1,4 @@
-import {makeLink, setup} from '@/__tests__/utils'
+import {makeLink, setup} from '@/__tests__/__helpers__'
 import {AddLinkForm} from '@/app/admin/components'
 import {Link} from '@/models'
 import {faker} from '@faker-js/faker'
