+++ src/__tests__/app/admin/components/header/index.test.tsx
@@ -1,8 +1,8 @@
 import * as firebaseAuth from 'firebase/auth'
 import routerMock from 'next-router-mock'
 
-import {fail, setup} from '@/__tests__/utils'
-import {makeUser} from '@/__tests__/utils/mocks'
+import {fail, setup} from '@/__tests__/__helpers__'
+import {makeUser} from '@/__tests__/__helpers__'
 import {Header} from '@/app/admin/components/header'
 import {authStore} from '@/app/auth/context/auth-store'
 import {User} from '@/models'
