+++ src/__tests__/app/components/smartphone/index.test.tsx
@@ -1,4 +1,4 @@
-import {setup} from '@/__tests__/utils'
+import {setup} from '@/__tests__/__helpers__'
 import {Smartphone} from '@/app/components'
 import {faker} from '@faker-js/faker'
